+++ data.json
@@ -1,0 +1,87 @@
+{
+  "title": "자바스크립트 월드컵 (16강)",
+  "count": 0,
+
+  "list": [
+    {
+      "id": 1,
+      "candidate": "1번",
+      "score": 0
+    },
+    {
+      "id": 2,
+      "candidate": "2번",
+      "score": 0
+    },
+    {
+      "id": 3,
+      "candidate": "3번",
+      "score": 0
+    },
+    {
+      "id": 4,
+      "candidate": "4번",
+      "score": 0
+    },
+    {
+      "id": 5,
+      "candidate": "5번",
+      "score": 0
+    },
+    {
+      "id": 6,
+      "candidate": "6번",
+      "score": 0
+    },
+    {
+      "id": 7,
+      "candidate": "7번",
+      "score": 0
+    },
+    {
+      "id": 8,
+      "candidate": "8번",
+      "score": 0
+    },
+    {
+      "id": 9,
+      "candidate": "9번",
+      "score": 0
+    },
+    {
+      "id": 10,
+      "candidate": "10번",
+      "score": 0
+    },
+    {
+      "id": 11,
+      "candidate": "11번",
+      "score": 0
+    },
+    {
+      "id": 12,
+      "candidate": "12번",
+      "score": 0
+    },
+    {
+      "id": 13,
+      "candidate": "13번",
+      "score": 0
+    },
+    {
+      "id": 14,
+      "candidate": "14번",
+      "score": 0
+    },
+    {
+      "id": 15,
+      "candidate": "15번",
+      "score": 0
+    },
+    {
+      "id": 16,
+      "candidate": "16번",
+      "score": 0
+    }
+  ]
+}
