+++ db.json
@@ -1,0 +1,178 @@
+{
+  "world": [
+    {
+      "title": "자바스크립트 월드컵 (16강)",
+      "count": 0,
+      "list": [
+        {
+          "id": 1,
+          "candidate": "1번",
+          "score": 0
+        },
+        {
+          "id": 2,
+          "candidate": "2번",
+          "score": 0
+        },
+        {
+          "id": 3,
+          "candidate": "3번",
+          "score": 0
+        },
+        {
+          "id": 4,
+          "candidate": "4번",
+          "score": 0
+        },
+        {
+          "id": 5,
+          "candidate": "5번",
+          "score": 0
+        },
+        {
+          "id": 6,
+          "candidate": "6번",
+          "score": 0
+        },
+        {
+          "id": 7,
+          "candidate": "7번",
+          "score": 0
+        },
+        {
+          "id": 8,
+          "candidate": "8번",
+          "score": 0
+        },
+        {
+          "id": 9,
+          "candidate": "9번",
+          "score": 0
+        },
+        {
+          "id": 10,
+          "candidate": "10번",
+          "score": 0
+        },
+        {
+          "id": 11,
+          "candidate": "11번",
+          "score": 0
+        },
+        {
+          "id": 12,
+          "candidate": "12번",
+          "score": 0
+        },
+        {
+          "id": 13,
+          "candidate": "13번",
+          "score": 0
+        },
+        {
+          "id": 14,
+          "candidate": "14번",
+          "score": 0
+        },
+        {
+          "id": 15,
+          "candidate": "15번",
+          "score": 0
+        },
+        {
+          "id": 16,
+          "candidate": "16번",
+          "score": 0
+        }
+      ]
+    },
+    {
+      "title": "asdasd",
+      "list": [
+        {
+          "id": "ba1d7d51-5752-4c6e-8f32-2de8fe658677",
+          "candidate": "xvz",
+          "score": 0
+        },
+        {
+          "id": "2138c601-5130-4070-af43-1867b6e8d772",
+          "candidate": "zxvzxv",
+          "score": 0
+        },
+        {
+          "id": "d52e49c6-9487-4146-a88e-07563918b92d",
+          "candidate": "xzvzxv",
+          "score": 0
+        },
+        {
+          "id": "a48faae2-c8e6-43ee-b2b0-c45e48d0dab9",
+          "candidate": "vzxv",
+          "score": 0
+        },
+        {
+          "id": "a2f223f7-f49a-4114-9949-d3bed5be5525",
+          "candidate": "xvzxv",
+          "score": 0
+        },
+        {
+          "id": "9a2928ab-d714-49bc-a6bb-a3ba1dd8d47f",
+          "candidate": "xvzxv",
+          "score": 0
+        },
+        {
+          "id": "bfc14b45-0b21-4b32-bc98-26951ece7da9",
+          "candidate": "zxvxv",
+          "score": 0
+        },
+        {
+          "id": "0867bcf6-5e2c-4e4a-94ee-4fc340f57fa9",
+          "candidate": "vzx",
+          "score": 0
+        },
+        {
+          "id": "f9fba09a-8862-40bf-89ea-e1d2803f8e7b",
+          "candidate": "asdas",
+          "score": 0
+        },
+        {
+          "id": "0f9c45da-feb4-4a14-8838-cda99972457f",
+          "candidate": "vzxvz",
+          "score": 0
+        },
+        {
+          "id": "bfe165db-d70e-463d-a702-a6e581d67ec4",
+          "candidate": "xvzx",
+          "score": 0
+        },
+        {
+          "id": "7e57699d-c607-46c5-b0b1-c1c155a1e7d0",
+          "candidate": "zxvzx",
+          "score": 0
+        },
+        {
+          "id": "4fd8e7b2-a37e-4a52-a095-f979bca3ad68",
+          "candidate": "fzxxz",
+          "score": 0
+        },
+        {
+          "id": "3e2d6796-e345-4a5e-b3fd-6159bcaddf17",
+          "candidate": "vzxvzx",
+          "score": 0
+        },
+        {
+          "id": "60041912-6b06-463d-8e95-c9e53db93b17",
+          "candidate": "vzxv",
+          "score": 0
+        },
+        {
+          "id": "b7bb194a-c446-4638-abb4-7760239a9582",
+          "candidate": "xzvzx",
+          "score": 0
+        }
+      ],
+      "id": "12814984-4e17-46c0-84d1-604993f33fc1",
+      "count": 0,
+      "createdAt": 1646415035229
+    }
+  ]
+}
