+++ src/index.tsx
@@ -1,8 +1,8 @@
+import App from 'App';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { ThemeProvider } from 'styled-components';
 import { theme, GlobalStyle } from 'styles/index';
-import App from './WorldCup';
 
 ReactDOM.render(
   <React.StrictMode>
