+++ src/pages/home/Home.tsx
@@ -1,6 +1,22 @@
-import React from 'react';
+import axios from 'axios';
+import React, { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 
 export default function Home() {
-  return <div>Home</div>;
+  const [listArr, setListArr] = useState<any>([{}]);
+  useEffect(() => {
+    axios.get('http://localhost:4000/world').then((res) => {
+      console.log(res);
+      setListArr(res.data);
+    });
+  }, []);
+  return (
+    <div>
+      {listArr.map((item: any) => (
+        <div key={item.id}>
+          <Link to={`/world/${item.id}`}>{item.title}</Link>
+        </div>
+      ))}
+    </div>
+  );
 }
